+++ server/src/main/java/br/com/dbserver/votacao/service/RestauranteService.java
@@ -6,35 +6,42 @@ import java.time.temporal.TemporalAdjusters;
 import java.util.List;
 import java.util.Optional;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import br.com.dbserver.votacao.exception.ResourceNotFoundException;
 import br.com.dbserver.votacao.model.Restaurante;
 import br.com.dbserver.votacao.repository.RestauranteRepository;
+import lombok.RequiredArgsConstructor;
 
 @Service
-public class RestauranteService 
+@RequiredArgsConstructor
+public class RestauranteService
 {
-    @Autowired
-    private RestauranteRepository repository;
+    private final RestauranteRepository repository;
     
     public List<Restaurante> findAll() 
     {
         return repository.findAll();
     }
     
-    public Optional<Restaurante> findById( Long id ) 
+    public Restaurante findById( Long id )
     {
-        return repository.findById( id );
+        return repository.findById( id )
+                         .orElseThrow( () -> new ResourceNotFoundException( "Restaurante não encontrado com ID: " + id ) );
     }
     
-    public Restaurante save( Restaurante restaurante ) 
+    public Restaurante save( Restaurante restaurante )
     {
         return repository.save( restaurante );
     }
     
-    public void delete( Long id ) 
+    public void delete( Long id )
     {
+        if ( ! repository.existsById( id ) ) 
+        {
+            throw new ResourceNotFoundException( "Restaurante não encontrado com ID: " + id );
+        }
+
         repository.deleteById( id );
     }
